refactor: StarRating 별점 Stasfactions 만족도에서 분리
