+++ src/main/java/com/backend/arthere/satisfactions/domain/SatisfactionsCustomRepository.java
@@ -13,5 +13,5 @@ public interface SatisfactionsCustomRepository {
 
     List<SatisfactionType> findSatisfactionsType(Long artId, Long userId);
 
-    Integer findStarRatings(Long artId, Long userId);
+
 }
