+++ src/main/java/com/backend/arthere/starRatings/domain/StarRatingsCustomRepository.java
@@ -1,0 +1,6 @@
+package com.backend.arthere.starRatings.domain;
+
+public interface StarRatingsCustomRepository {
+
+    Integer findStarRatings(Long artId, Long userId);
+}
