+++ src/main/java/com/backend/arthere/starRatings/domain/StarRatingsRepository.java
@@ -1,0 +1,6 @@
+package com.backend.arthere.starRatings.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StarRatingsRepository extends JpaRepository<StarRatings, Long>, StarRatingsCustomRepository {
+}
